--- a/app/models/edition.rb
+++ b/app/models/edition.rb
@@ -36,11 +36,13 @@ class Edition < ApplicationRecord
 
   belongs_to :document
   has_one :unpublishing
+  has_one :change_note
   has_many :links
 
   scope :renderable_content, -> { where.not(document_type: NON_RENDERABLE_FORMATS) }
   scope :with_document, -> { joins(:document) }
   scope :with_unpublishing, -> { left_outer_joins(:unpublishing) }
+  scope :with_change_note, -> { left_outer_joins(:change_note) }
 
   validates :document, presence: true