This is to try reduce the Specialist Publisher timeouts that are a pain /cc @NeilvB
It's all a bit ugly and messy, suggestions to clean up the code (or even better speed up the queries are very welcome)

tl;dr

Benchmarks between this and fcd37be show a speed up in accessing the
slower Specialist Publisher index pages

from:

```
dfid_research_output
..........  5.620000   0.020000   5.640000 ( 49.909101)

aaib_report
..........  1.900000   0.000000   1.900000 ( 18.333954)
```

to:

```
dfid_research_output
..........  0.010000   0.010000   0.020000 (  6.626270)

aaib_report
..........  0.020000   0.000000   0.020000 (  2.608745)
```

Full benchmarks at the end.

anyway...

This changes the way the ContentItemPresenter works so that it no longer
pulls out a large collection of ids that can be queried and instead does
a single query that pulls out everything, while making use of state to
reduce data set size.

To group by latest items in the query we use `DISTINCT ON` which
unfortunately means that to do ordering we need to run that as a
subquery inside a `SELECT * FROM subquery`.

This can give us some nice performance gains for a document_type that
has many editions - although the query is 4 or 5 times faster than the
previous iteration it still is quite slow, in the region of 600ms.

Digging into the presenters/queries/content_item_presenter class is
particularly difficult, and it looks like it's got way over complicated
and is used in too many cases. This trello card:
https://trello.com/c/m7PFbmGt/399-remove-state-history-from-get-content-and-rethink-endpoint
suggests steps.

Full benchmarks:

master:
```
vm  publishing-api git:(fcd37be) ruby bench/get_content_items.rb
..........  0.640000   0.250000   0.890000 (  2.645056)
queries: 30

vm  publishing-api git:(fcd37be) ruby bench/content_item_presenter.rb
Presenting 64 content items
  0.030000   0.010000   0.040000 (  0.078049)
  Queries: 4
Presenting 128 content items
  0.000000   0.000000   0.000000 (  0.020408)
  Queries: 2
Presenting 256 content items
  0.020000   0.010000   0.030000 (  0.041716)
  Queries: 2
Presenting 512 content items
  0.030000   0.000000   0.030000 (  0.077993)
  Queries: 2
Presenting 1024 content items
  0.080000   0.010000   0.090000 (  0.152250)
  Queries: 2
vm  publishing-api git:(fcd37be) ruby bench/specialist_publisher_index_content.rb
countryside_stewardship_grant
..........  0.090000   0.020000   0.110000 (  0.378548)

gone
..........  0.020000   0.000000   0.020000 (  0.224220)

redirect
..........  0.160000   0.010000   0.170000 (  2.241235)

esi_fund
..........  0.110000   0.000000   0.110000 (  0.487870)

international_development_fund
..........  0.040000   0.000000   0.040000 (  0.245753)

service_standard_report
..........  0.060000   0.000000   0.060000 (  0.287343)

medical_safety_alert
..........  0.110000   0.020000   0.130000 (  0.463453)

specialist_document
..........  0.120000   0.000000   0.120000 (  0.489946)

employment_appeal_tribunal_decision
..........  0.020000   0.010000   0.030000 (  0.199861)

finder_email_signup
..........  0.020000   0.000000   0.020000 (  0.215405)

finder
..........  0.030000   0.000000   0.030000 (  0.220461)

vehicle_recalls_and_faults_alert
..........  0.030000   0.000000   0.030000 (  0.208663)

asylum_support_decision
..........  0.030000   0.000000   0.030000 (  0.229216)

dfid_research_output
..........  5.620000   0.020000   5.640000 ( 49.909101)

aaib_report
..........  1.900000   0.000000   1.900000 ( 18.333954)

raib_report
..........  0.120000   0.000000   0.120000 (  0.429986)

cma_case
..........  0.340000   0.010000   0.350000 (  1.843951)

drug_safety_update
..........  0.110000   0.010000   0.120000 (  0.445271)

utaac_decision
..........  0.110000   0.000000   0.110000 (  0.368779)

maib_report
..........  0.180000   0.000000   0.180000 (  0.966065)

tax_tribunal_decision
..........  0.140000   0.010000   0.150000 (  0.697152)

employment_tribunal_decision
..........  0.100000   0.000000   0.100000 (  0.383149)
```

this branch:
```
vm  publishing-api git:(refactor-get-content) ruby bench/get_content_items.rb
..........  0.320000   0.240000   0.560000 (  2.722570)
queries: 19

vm  publishing-api git:(refactor-get-content) ruby bench/content_item_presenter.rb
Presenting 64 content items
  0.020000   0.010000   0.030000 (  0.048117)
  Queries: 3
Presenting 128 content items
  0.000000   0.000000   0.000000 (  0.007047)
  Queries: 1
Presenting 256 content items
  0.000000   0.000000   0.000000 (  0.007677)
  Queries: 1
Presenting 512 content items
  0.000000   0.000000   0.000000 (  0.010066)
  Queries: 1
Presenting 1024 content items
  0.010000   0.000000   0.010000 (  0.024324)
  Queries: 1
vm  publishing-api git:(refactor-get-content) ruby bench/specialist_publisher_index_content.rb
countryside_stewardship_grant
..........  0.030000   0.020000   0.050000 (  0.301340)

gone
..........  0.010000   0.000000   0.010000 (  0.202118)

redirect
..........  0.030000   0.000000   0.030000 (  0.553600)

esi_fund
..........  0.020000   0.000000   0.020000 (  0.309609)

international_development_fund
..........  0.020000   0.000000   0.020000 (  0.209377)

service_standard_report
..........  0.020000   0.010000   0.030000 (  0.226802)

medical_safety_alert
..........  0.010000   0.000000   0.010000 (  0.289368)

specialist_document
..........  0.020000   0.000000   0.020000 (  0.208020)

employment_appeal_tribunal_decision
..........  0.010000   0.010000   0.020000 (  0.192036)

finder_email_signup
..........  0.020000   0.000000   0.020000 (  0.208976)

finder
..........  0.020000   0.010000   0.030000 (  0.203342)

vehicle_recalls_and_faults_alert
..........  0.010000   0.000000   0.010000 (  0.195428)

asylum_support_decision
..........  0.020000   0.000000   0.020000 (  0.205268)

dfid_research_output
..........  0.010000   0.010000   0.020000 (  6.626270)

aaib_report
..........  0.020000   0.000000   0.020000 (  2.608745)

raib_report
..........  0.010000   0.000000   0.010000 (  0.275698)

cma_case
..........  0.020000   0.010000   0.030000 (  0.587719)

drug_safety_update
..........  0.010000   0.000000   0.010000 (  0.295200)

utaac_decision
..........  0.030000   0.000000   0.030000 (  0.258562)

maib_report
..........  0.020000   0.000000   0.020000 (  0.393338)

tax_tribunal_decision
..........  0.020000   0.000000   0.020000 (  0.308831)

employment_tribunal_decision
..........  0.020000   0.000000   0.020000 (  0.245689)
```