@kevindew 2017-03-03
title edited after @carlosmartinez pointed out refactor was a touch unfitting for such a mess 😉