@kevindew 2017-03-03
@thomasleese I've updated it to be two methods